Updated to use player uuids. UUID for shops get stored as soon as the player logs in. The playername gets used for display purposes and as identifier for the owner as long as we don't have an owner uuid yet.
Caveat which is probably left: the 'remove shopkeeper after certain time
of inactivity of player' gets an OfflinePlayer to determine when the
player was last online. This however can now (since 1.7.8) freeze the
server when it has to lookup the name or uuid via the mojang services.
However, this is currently considered a minor 'issue' as those lookups
can only happen once per plugin load (and reload), and the server might
cache the result from those lookups.