The ShopCreationData is now passed through to the ShopObject and used to provide the NPC id for Shopkeepers created by the CitizensShopkeeperTrait.
Fixed some endless-loop between removing a CitizensShopkeeperTrait and
the removing of the Shopkeeper. Also Citizens NPC's with the
CitizensShopkeeperTrait are no longer removed when the linked Shopkeeper
or this trait get removed.

Probably fixed some duplicated Citizens NPC creation.

The created Citizens NPC should now get spawned on creation.

Citizens NPC' of PlayerShopkeepers should now default to Player-type
with the name of the owner.

Untested yet.