Additions to last commit: More fixes regarding owner uuids and running on older bukkit versions. (getOwner())
Another change:
We now no longer delete the owner uuid during a shopkeeper save in case
the bukkit version we are running on does not properly support player
uuids yet. Instead we keep storing the unused uuid, known from running
on bukkit versions which supported player uuids. This is useful in case
the server switches back to a bukkit version supporting owner uuids
again. However, we need to make sure that no uuids are getting stored,
which originate from running on a bukkit version which doesn't support
player uuids properly yet. (the only case for this was during shopkeeper
creation, which got removed now with this commit)