Updated to minecraft's new trading mechanics in 1.8: mc is now comparing for exact quality, if the required item has special data.
Added use-strict-item-comparison setting (default true, in order to not
cause problems if running on mc versions below 1.8), which allows
turning off our own item comparison during trades.
So if set to false and minecraft allows a trade, we usually try to allow
it as well then.

If set to true we additionally make sure that the items offered and the
items required are perfectly matching, and if they are not we cancle the
trade. This mostly matches previous behavior.

Removed our own item comparison again, which is now useless, because
minecraft itself is now (Mc 1.8) already performing a strict comparison
for items with custom data, resulting in the problems we tried to fix
with our own item comparison (comparing attributes, ignoring uuids).
Also removed related code from the different modules.

Some more refactoring and small fixes related to the latest commits.

Not much tested yet.