Possible workaround for duplicating shopkeeper entities.
The problem (my current guess..):
* The shopkeeper entities get saved to disk when the world gets saved.
* Shopkeeper removes a shop entity, respawns an entity (this replaces
the old 'last-entity-uuid'), and removes it again. This does not
necessarily have to happen at the same time or for all entities, but it
does occur when shopkeepers is for ex. getting reloaded.
* The server closes (ex. due to a crash) without another world save
taking place.

The old shop entity is still saved in the chunk data of the world, but
we cannot find it on the next server start, because we remembered the
entity-uuid of the last entity we spawned, which is not the same as the
last entity which was saved to disk.

This workaround:
Whenever the world gets saved (whenever WorldSaveEvent gets triggered)
all shopkeepers get unloaded and loaded again shortly after the
world-save.