Added setTradePerm command (and other small changes).
This allows setting of a custom permission for individual admin
shopkeepers, which is required for being allowed to trade with the shop.

Added a few new messages.

The 'remote' command now also accepts a shop's unique id or session id
(instead of the shop name). The default 'unknown-shopkeeper'-message was
changed to include the 'unknown id' fact.

Also added a new message which gets sent to players which don't have the
regular trade-permission. By default the message text is the same as the
message being sent if the player is missing the newly added custom
trading permission, though they can be changed in the config
independently.