Removed static LivingEntityType enum and did some refactoring.
The LivingEntityType enum was replaced with LivingEntityObjectTypes,
which dynamically creates the living object types for ALL living entity
types available to the server.
Instead of having the selection order defined statically by the order of
the enum, it is now defined by the order of the entries of
enabled-living-shops setting in the config. So by default the villager
object type is still the first in the shop object selection cycling,
however that can now be changed in the config.

Internal refactoring related to DefaultShopObjectTypes and
DefaultShopTypes:
Instead of statically defining the default shop and object types, they
now get defined freshly every time the plugin gets enabled/reloaded.
This has the advantage, that those can now more easily rely on config
settings during initialization (like the living entity shop object types
relying on the order of the values for the enabled-living-types setting
now).

Also created separate class files for the default shop and object types
(instead of having them anonymously in the same class).
Moved some common shop type methods into the ShopType class.