Additionally check to ignore shopkeeper citizens trait attachment, if shopkeepers isn't currently running.
Usually this shouldn't get triggered anymore though, due to the changes
of the last commit.

Whenever citizens npc's with shopkeeper attachment get reloaded and
shopkeepers is running, onAttach is called and it is checked if there is
already a shopkeeper setup for that shopkeeper trait. If not it freshly
creates one.