We changed the way we open the trading window in somewhere in 1.8.
With this change it might no longer be an problem to open the trading
menu even if the player is holding a villager spawn egg in hand.

TODO: Actually test this, and then remove the current restriction again.
TODO: Also test with 1.7R4, which currently still uses the old method of
opening the trading menu, but might be able to get updated to the new
method.