Reimplemented chest protection check:
Now it considers all shopkeepers, even currently inactive ones. This
might be of use in the edge case of player shopkeepers being in a
different and somehow unloaded chunk than the corresponding chest.
This should also fix the issue that chest are temporarily unprotected
after world saves due to the temporarily deactivation of shops.
And this might bring slight performance improvements because it now uses
hashing instead of looping through all active shopkeepers.