A few minor changes, mostly related to last few commits:
Added a new setting 'naming-of-player-shops-via-item' which can even be
used independently of the new chest-option-setting to disable the normal
naming button and only allow naming via item. When the chest-option is
enabled, this has to be enabled as well in order for players to be still
able to name their shops.
The naming item can now require a specific display name and lore.
Shopkeepers without name will stores this as an empty string instead of
null now.
Moved naming procedure into Shopkeeper.
A few other minor changes.