Update for MC 1.11:
* Minecraft represents the different variants of skeletons (stray,
wither_skeleton) and zombies (zombie_villager) now in the form of
different entity types.
To save me some work and trouble I decided to not try to convert the
shopkeepers using the previous skeleton and zombie variants to new
shopkeepers using the new mob types. So shopkeepers using those variants
will fallback to normal skeletons and zombies once loaded with this
version. You will have to recreate your affected shops if you want to
use specific zombie or skeleton variants as shopkeepers.
Or alternatively, before you run this version, manually convert the
affected shopkeepers in the save.yml file: For that, stop the server,
make a backup and then search for entries for which the value for
'skeletonType' is not 'NORMAL', and entries for which 'villagerZombie'
is 'true'. For those entries you will have to change the value of
'object' to the correct mob id ('stray', 'wither_skeleton' or
'zombie_villager').

* Added a bunch of old mobs, which seem to mostly work fine, to the
default enabled living shop types. Also reordered them a bit. And added
the new mobs from MC 1.11 which don't seem to cause severe issues.
LLamas, sadly, act weirdly currently when clicked (same goes for all
horse variants) and are therefore not included in the default list of
enabled mob types.

* Added: Zombie shopkeepers can now change to baby variant by clicking
the spawn egg in the editor window.

* Some minor changes related to handling empty inventory slots and
itemstacks in inventories and trading recipes were required.